chore(lint): lint all script, modify makefile python source

+++ scripts/update_dependency.sh
@@ -1,9 +1,14 @@
 #!/bin/sh
 
+if [ $# -eq 0 ]; then
+  echo "No pip supplied"
+  exit 1
+fi
+
 pip=$1
 
-$pip freeze > requirements.txt && \
-$pip uninstall -y -r requirements.txt
+$pip freeze >requirements.txt &&
+  $pip uninstall -y -r requirements.txt
 rm -rf requirements.txt
 
 $pip install --upgrade pip
@@ -22,6 +27,6 @@ $pip install -U \
   Jinja2 \
   prometheus-client
 
-$pip freeze > requirements.txt
+$pip freeze >requirements.txt
 
 git add requirements.txt
